+++ src/mailer/base.js
@@ -52,43 +52,6 @@ export class VelociousMailerBase {
     this._mailOptions = null
     this._viewParams = {}
     this._configurationPromise = configuration ? Promise.resolve(configuration) : configurationResolver()
-    this._actionMethods = new Map()
-    this._setupActionWrappers()
-  }
-
-  /**
-   * @returns {void} - No return value.
-   */
-  _setupActionWrappers() {
-    const basePrototype = VelociousMailerBase.prototype
-    const wrappedNames = new Set()
-    let prototype = Object.getPrototypeOf(this)
-
-    while (prototype && prototype !== basePrototype) {
-      for (const name of Object.getOwnPropertyNames(prototype)) {
-        if (name === "constructor" || wrappedNames.has(name)) continue
-
-        const descriptor = Object.getOwnPropertyDescriptor(prototype, name)
-
-        if (!descriptor || typeof descriptor.value !== "function") continue
-        if (name.startsWith("_") || name in basePrototype) continue
-
-        wrappedNames.add(name)
-        this._actionMethods.set(name, descriptor.value)
-
-        this[name] = (...args) => this._buildDelivery(name, args)
-      }
-
-      prototype = Object.getPrototypeOf(prototype)
-    }
-  }
-
-  /**
-   * @param {string} actionName - Action name.
-   * @returns {((...args: any[]) => unknown) | undefined} - Action method.
-   */
-  _getActionMethod(actionName) {
-    return this._actionMethods.get(actionName)
   }
 
   /**
@@ -108,32 +71,24 @@ export class VelociousMailerBase {
    * @param {any} [args.bcc] - BCC recipients.
    * @param {any} [args.replyTo] - Reply-to address.
    * @param {Record<string, string>} [args.headers] - Custom headers.
-   * @returns {void} - No return value.
-   */
-  mail({to, subject, from, cc, bcc, replyTo, headers, ...restArgs}) {
-    restArgsError(restArgs)
-
-    this._mailOptions = {to, subject, from, cc, bcc, replyTo, headers}
-  }
-
-  /**
-   * @param {string} actionName - Action name.
-   * @param {any[]} args - Action args.
+   * @param {string} args.actionName - Mailer action name.
+   * @param {Promise<unknown> | unknown} [args.actionPromise] - Action completion promise.
    * @returns {MailerDelivery} - Delivery wrapper.
    */
-  _buildDelivery(actionName, args) {
-    const action = this._getActionMethod(actionName)
+  mail({to, subject, from, cc, bcc, replyTo, headers, actionName, actionPromise, ...restArgs}) {
+    restArgsError(restArgs)
 
-    if (typeof action !== "function") {
-      throw new Error(`Unknown mailer method "${actionName}" on ${this.constructor.name}`)
+    if (!actionName) {
+      throw new Error(`Missing actionName for ${this.constructor.name}.mail()`)
     }
 
     this._actionName = actionName
-    const actionResult = action.apply(this, args)
+    this._mailOptions = {to, subject, from, cc, bcc, replyTo, headers}
+    const resolvedActionPromise = actionPromise === undefined ? Promise.resolve() : Promise.resolve(actionPromise)
 
     return new MailerDelivery({
       mailer: this,
-      actionPromise: Promise.resolve(actionResult),
+      actionPromise: resolvedActionPromise,
       actionName
     })
   }